SignallingNamespace, update Express-session, add method for find user by socket id

+++ backend/src/socket/namespaces/game/game-namespace.ts
@@ -6,23 +6,36 @@ import gameService from "../../../services/game-service";
  * */
 export default ((io: Namespace) => {
 	io.on('connection', (socket) => {
-	
-		const userId = socket.id;
-		
-		gameService.addUser(userId);
+
+		const userId = socket.request.session?.userId;
+		const socketId = socket.id;
+		if (!userId) return socket.disconnect();
+
+		gameService.addUser(userId, socketId);
 		
 		/**
 		 * @description User create event 'new' when want to find a new partner.
 		 * */
-		socket.on('new', async () => {
+		socket.on('partner:new', async () => {
 			const freeUser = await gameService.findFree(userId)
 			
 			if (!freeUser) return;
 
-			//gameService.matchUsers(userId, freeUser.userId);
+			const partnerId = freeUser.userId;
 
+			await gameService.matchUsers(userId, partnerId);
+			const partnerSocket = await gameService.getSocketIdByUserId(partnerId);
+
+			io.sockets.get(partnerSocket)?.emit('connect-to', socketId);
+			socket.emit('connect-to', partnerId)
 		})
-		
+
+		socket.on('signalling', async () => {});
+
+		/**
+		 * 1. Перед удаленеим, отключить партнёров
+		 * 2. Удаление всей инфорамции из таблицы
+		 */
 		socket.on('disconnect', () => {
 			gameService.removeUser(userId);
 		})
